+++ frappe_slack_connector/api/slack_interactions.py
@@ -6,21 +6,12 @@ from frappe import _
 from frappe_slack_connector.helpers.error import generate_error_log
 from frappe_slack_connector.helpers.http_response import send_http_response
 from frappe_slack_connector.slack.app import SlackIntegration
-from frappe_slack_connector.slack.interactions.approve_leave import (
-    handler as approve_leave_handler,
-)
-from frappe_slack_connector.slack.interactions.submit_leave import (
-    half_day_checkbox_handler,
-)
-from frappe_slack_connector.slack.interactions.submit_leave import (
-    handler as submit_leave_handler,
-)
-from frappe_slack_connector.slack.interactions.submit_timesheet import (
-    handler as submit_timesheet_handler,
-)
-from frappe_slack_connector.slack.interactions.timesheet_filters import (
-    handle_timesheet_filter,
-)
+from frappe_slack_connector.slack.interactions.approve_leave import handler as approve_leave_handler
+from frappe_slack_connector.slack.interactions.submit_leave import half_day_checkbox_handler
+from frappe_slack_connector.slack.interactions.submit_leave import handler as submit_leave_handler
+from frappe_slack_connector.slack.interactions.submit_timesheet import handler as submit_timesheet_handler
+from frappe_slack_connector.slack.interactions.timesheet_filters import handle_timesheet_filter
+from frappe_slack_connector.slack.interactions.timesheet_modal import show_timesheet_modal
 
 
 @frappe.whitelist(allow_guest=True)
@@ -72,6 +63,8 @@ def event():
             # action payload is not required
             if action_id.startswith("ignore"):
                 return
+            elif block_id == "daily_reminder_button":
+                return show_timesheet_modal(slack, payload["user"]["id"], payload["trigger_id"])
             elif block_id == "half_day_checkbox":
                 return half_day_checkbox_handler(slack, payload)
             elif block_id in ("project_block", "task_block"):
@@ -97,6 +90,4 @@ def event():
 
     except Exception as e:
         generate_error_log("Error handling the event", exception=e)
-        frappe.throw(
-            _("An error occurred while handling the event"), frappe.PermissionError
-        )
+        frappe.throw(_("An error occurred while handling the event"), frappe.PermissionError)
